Hash uses Hyatt's lockless hashing now.

+++ src/eval.h
@@ -10,7 +10,7 @@ const int maxMateScore = 32767 - maxGameLength;
 const int infinity = mateScore + 1;
 
 // Watch out with Infinite as it is bigger than mateScore therefore it could be detected as an mate score even though it is not.
-inline bool isMateScore(int score)
+inline bool isMateScore(int64_t score)
 {
 	return (score <= -maxMateScore || score >= maxMateScore);
 }
